update approach to footer buttons
