+++ ts/components/RemindEmailValidationOverlay.tsx
@@ -35,6 +35,7 @@ import customVariables from "../theme/variables";
 import TopScreenComponent, {
   TopScreenComponentProps
 } from "./screens/TopScreenComponent";
+import BlockButtons from "./ui/BlockButtons";
 import FooterWithButtons from "./ui/FooterWithButtons";
 import IconFont from "./ui/IconFont";
 import Markdown from "./ui/Markdown";
@@ -265,39 +266,46 @@ class RemindEmailValidationOverlay extends React.PureComponent<Props, State> {
     // show two buttons where the left one is a CTA
     // to edit again the email
     return (
-      <FooterWithButtons
-        type={"TwoButtonsInlineThirdInverted"}
-        upperButton={{
-          title: this.state.ctaSendEmailValidationText,
-          onPress: this.handleSendEmailValidationButton,
-          light: true,
-          bordered: true,
-          disabled:
-            this.state.isLoading || this.state.isCtaSentEmailValidationDisabled
-        }}
-        leftButton={{
-          block: true,
-          bordered: true,
-          disabled: this.state.isLoading,
-          onPress: () => {
-            if (!isOnboardingCompleted) {
-              this.props.closeModalAndNavigateToEmailInsertScreen();
-              return;
-            }
-            this.props.navigateToEmailInsertScreen();
-          },
-          title: I18n.t("email.edit.title")
-        }}
-        rightButton={{
-          block: true,
-          primary: true,
-          onPress: this.handleOnClose,
-          disabled: this.state.isLoading,
-          title: isOnboardingCompleted
-            ? I18n.t("global.buttons.ok")
-            : I18n.t("global.buttons.continue")
-        }}
-      />
+      <View footer={true}>
+        <BlockButtons
+          type={"SingleButton"}
+          leftButton={{
+            title: this.state.ctaSendEmailValidationText,
+            onPress: this.handleSendEmailValidationButton,
+            light: true,
+            bordered: true,
+            disabled:
+              this.state.isLoading ||
+              this.state.isCtaSentEmailValidationDisabled
+          }}
+        />
+        <View spacer={true} />
+        <BlockButtons
+          type={"TwoButtonsInlineThirdInverted"}
+          leftButton={{
+            block: true,
+            bordered: true,
+            disabled: this.state.isLoading,
+            onPress: () => {
+              if (!isOnboardingCompleted) {
+                this.props.closeModalAndNavigateToEmailInsertScreen();
+                return;
+              }
+              this.props.navigateToEmailInsertScreen();
+            },
+            title: I18n.t("email.edit.title")
+          }}
+          rightButton={{
+            block: true,
+            primary: true,
+            onPress: this.handleOnClose,
+            disabled: this.state.isLoading,
+            title: isOnboardingCompleted
+              ? I18n.t("global.buttons.ok")
+              : I18n.t("global.buttons.continue")
+          }}
+        />
+      </View>
     );
   };
