+++ ts/theme/components/Button.ts
@@ -16,21 +16,9 @@ declare module "native-base" {
 
 export default (): Theme => {
   return {
-    ".block": {
-      ".iconVeryLeft": {
-        "UIComponent.IconFont": {
-          flex: 0,
-          borderRightWidth: 1,
-          borderColor: "#FFFFFF",
-          margin: 0,
-          padding: 11,
-          paddingLeft: 15,
-          paddingRight: 15
-        },
-        "NativeBase.Text": { flex: 1, textAlign: "center" },
-        padding: 0,
-        display: "flex"
-      }
+    ".info": {
+      backgroundColor: variables.brandHighLighter,
+      bordercolor: variables.brandHighLighter
     },
 
     ".xsmall": {
@@ -60,6 +48,9 @@ export default (): Theme => {
           color: variables.btnLightTextColor,
           fontWeight: variables.textBoldWeight
         },
+        "UIComponent.IconFont": {
+          color: variables.btnLightTextColor
+        },
         borderWidth: 1,
         borderColor: variables.btnLightBorderColor,
         backgroundColor: variables.brandLight,
@@ -70,7 +61,9 @@ export default (): Theme => {
           backgroundColor: variables.colorWhite
         }
       },
-      "NativeBase.Text": { fontWeight: variables.textNormalWeight }
+      "NativeBase.Text": {
+        fontWeight: variables.textNormalWeight
+      }
     },
 
     ".bordered": {
@@ -78,15 +71,24 @@ export default (): Theme => {
         "NativeBase.Icon": {
           color: variables.btnDisabledBg
         },
-
+        "UIComponent.IconFont": {
+          color: variables.btnDisabledBg
+        },
         backgroundColor: variables.colorWhite
+      },
+      "UIComponent.IconFont": {
+        color: variables.brandPrimary
       }
     },
 
     ".white": { backgroundColor: "#FFFFFF" },
+
     ".cancel": {
       backgroundColor: variables.brandDarkGray,
       "NativeBase.Text": {
+        color: variables.colorWhite
+      },
+      "UIComponent.IconFont": {
         color: variables.colorWhite
       }
     },
@@ -94,6 +96,10 @@ export default (): Theme => {
     "NativeBase.Text": {
       ...makeFontStyleObject(Platform.select, variables.btnTextFontWeight),
       fontSize: variables.btnFontSize
+    },
+
+    "UIComponent.IconFont": {
+      color: variables.colorWhite
     },
 
     justifyContent: "center",
